+++ internal/resolvers/twitch/data_test.go
@@ -1,0 +1,40 @@
+package twitch
+
+const goodSlugV1 = "GoodSlugV1"
+const goodSlugV2 = "GoodSlugV2-HVUvT7bYQnMn6nwp"
+
+var validClipBase = []string{
+	"https://clips.twitch.tv/",
+	"https://twitch.tv/pajlada/clip/",
+	"https://twitch.tv/zneix/clip/",
+	"https://m.twitch.tv/pajlada/clip/",
+	"https://m.twitch.tv/zneix/clip/",
+	"https://m.twitch.tv/clip/",
+	"https://m.twitch.tv/clip/clip/",
+}
+
+// clips that are invalid due to path or domain+path combination
+var invalidClips = []string{
+	"https://clips.twitch.tv/pajlada/clip/VastBitterVultureMau5",
+	"https://clips.twitch.tv/",
+	"https://twitch.tv/nam____________________________________________/clip/someSlugNam",
+	"https://twitch.tv/supinic/clip/",
+	"https://twitch.tv/pajlada/clips/VastBitterVultureMau5",
+	"https://twitch.tv/zneix/clip/ImpossibleOilyAlpacaTF2John-jIlgtnSAQ52BThHhifyouseethisvivon",
+	"https://twitch.tv/clip/slug",
+	"https://gql.twitch.tv/VastBitterVultureMau5",
+	"https://gql.twitch.tv/ThreeLetterAPI/clip/VastBitterVultureMau5",
+	"https://m.twitch.tv/VastBitterVultureMau5",
+	"https://m.twitch.tv/username/clip/clip/slug",
+	"https://m.twitch.tv/username/notclip/slug",
+}
+
+// clips that are invalid due to the path
+var invalidClipSlugs = []string{
+	"https://clips.twitch.tv/",
+	"https://twitch.tv/nam____________________________________________/clip/someSlugNam",
+	"https://twitch.tv/supinic/clip/",
+	"https://twitch.tv/pajlada/clips/VastBitterVultureMau5",
+	"https://twitch.tv/zneix/clip/ImpossibleOilyAlpacaTF2John-jIlgtnSAQ52BThHhifyouseethisvivon",
+	"https://m.twitch.tv/username/notclip/slug",
+}